Fix datatype issue with softmax
In some environments (e.g., conda) we see that the triton kernel source data types must exactly match the python datatype input. It turned out that in this case our number of blocks being passed was a torch.Tensor which when serialized by triton would lead to an exception. After ensuring that num_blocks was a python integer this seems to have fixed our issue.